battery-monitor/requirements.txt
mppsolar[ble]
python-telegram-bot[job-queue]
python-dotenv
pytz